refactor(viewer): reduce load_fea_streaming to orchestration and re-exports
The toggle entry point moves to fea/streaming/loadWithDefaults with the
load function injected, so the package never imports the loader it
feeds; the loader's argument type moves to fea/streaming/types so the
step callback and the defaults path can name it. load_fea_streaming.ts
is now the sequence of one load plus the facade every importer already
uses -- 1,815 lines to ~300 -- and holds no module-level state.
